fix: split font by environment switch
